requirements.txt
playwright
python-dotenv
playwright-stealth
